fixed annoying warnings

+++ vowpalwabbit/gd.cc
@@ -280,7 +280,7 @@ void print_features(vw& all, example& ec)
         for (audit_data* a = ns.begin; a != ns.end; ++a)
 	    {
             audit_interaction(dat, a);
-            audit_feature(dat, a->x, a->weight_index + ec.ft_offset);
+            audit_feature(dat, a->x, (uint32_t)a->weight_index + ec.ft_offset);
             audit_interaction(dat, NULL);
 	    }
 	}
@@ -382,7 +382,7 @@ void multipredict(gd& g, base_learner&, example& ec, size_t count, size_t step, 
       print_audit_features(all, ec);
       ec.ft_offset += (uint32_t)step;
     }
-    ec.ft_offset -= (uint32_t)step*count;
+    ec.ft_offset -= (uint32_t)(step*count);
   }
 }
